pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llamaindex-rag"
version = "0.1.0"
description = "Retrieval-Augmented Generation system using LlamaIndex"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "llama-index>=0.9.0",
    "llama-index-embeddings-openai>=0.1.3",
    "llama-index-llms-openai>=0.1.3",
    "llama-index-readers-file>=0.1.3",
    "llama-index-vector-stores-chroma>=0.1.1",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "streamlit>=1.28.0",
    "typer>=0.9.0",
    "rich>=13.6.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.10.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.1",
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/llamaindex-rag"
"Bug Reports" = "https://github.com/yourusername/llamaindex-rag/issues"
"Source" = "https://github.com/yourusername/llamaindex-rag"

[project.scripts]
llamarag = "llamaindex_rag.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=llamaindex_rag" 